Create public INes interface and cleanup exported API
This limits the exports we expose to a single Nes_Create function that
returns an INes pointer. INes is a public C++ interface that exposes Nes
functionality to the consumer of the nes library.

+++ src/nes.cpp
@@ -79,7 +79,7 @@ void Nes::DoFrame(u8 screen[])
     } while (!ppuResult.VBlank);
 }
 
-IStandardController* Nes::GetStandardController(u8 port)
+IStandardController* Nes::GetStandardController(unsigned int port)
 {
     return _input->GetStandardController(port);
 }
